rename shared cards -> common cards
