+++ fronts-client/src/bundles/frontsUIBundle.ts
@@ -40,7 +40,7 @@ import {
 import {
   REMOVE_GROUP_CARD,
   REMOVE_SUPPORTING_CARD
-} from 'shared/actions/Cards';
+} from 'actions/CardsCommon';
 import { Stages, CardSets } from 'types/Collection';
 import { selectPriority } from 'selectors/pathSelectors';
 import { CollectionWithArticles } from 'types/PageViewData';
