+++ fronts-client/src/keyboard/index.ts
@@ -11,7 +11,7 @@ import { ThunkResult } from 'types/Store';
 import Mousetrap from 'mousetrap';
 import { selectFocusState, setFocusState } from 'bundles/focusBundle';
 import { RefDrop } from 'util/collectionUtils';
-import { createArticleEntitiesFromDrop } from 'shared/actions/Cards';
+import { createArticleEntitiesFromDrop } from 'actions/CardsCommon';
 import { moveUp, moveDown } from './keyboardActionMaps/move';
 import { Card } from '../types/Collection';
 import { thunkInsertClipboardCard } from 'actions/Clipboard';
